+++ lib/global-styles-and-settings.php
@@ -347,10 +347,6 @@ function _gutenberg_clean_theme_json_caches() {
 	wp_cache_delete( 'gutenberg_get_global_stylesheet', 'theme_json' );
 	wp_cache_delete( 'gutenberg_get_global_settings_custom', 'theme_json' );
 	wp_cache_delete( 'gutenberg_get_global_settings_theme', 'theme_json' );
-	wp_cache_delete( 'gutenberg_get_global_styles_custom', 'theme_json' );
-	wp_cache_delete( 'gutenberg_get_global_styles_custom_resolved', 'theme_json' );
-	wp_cache_delete( 'gutenberg_get_global_styles_theme', 'theme_json' );
-	wp_cache_delete( 'gutenberg_get_global_styles_theme_resolved', 'theme_json' );
 	wp_cache_delete( 'gutenberg_get_global_custom_css', 'theme_json' );
 	wp_cache_delete( 'gutenberg_get_global_styles_base_custom_css', 'theme_json' );
 	WP_Theme_JSON_Resolver_Gutenberg::clean_cached_data();
@@ -397,21 +393,11 @@ function gutenberg_get_global_styles( $path = array(), $context = array() ) {
 	&& is_array( $context['transforms'] )
 	&& in_array( 'resolve-variables', $context['transforms'], true );
 
-	$cache_group = 'theme_json';
-	$cache_key   = 'gutenberg_get_global_styles_' . $origin;
+	$merged_data = WP_Theme_JSON_Resolver_Gutenberg::get_merged_data( $origin );
 	if ( $resolve_variables ) {
-		$cache_key .= '_resolved';
+		$merged_data = WP_Theme_JSON_Gutenberg::resolve_variables( $merged_data );
 	}
-	$styles = wp_cache_get( $cache_key, $cache_group );
-
-	if ( false === $styles || WP_DEBUG ) {
-		$merged_data = WP_Theme_JSON_Resolver_Gutenberg::get_merged_data( $origin );
-		if ( $resolve_variables ) {
-			$merged_data = WP_Theme_JSON_Gutenberg::resolve_variables( $merged_data );
-		}
-		$styles = $merged_data->get_raw_data()['styles'];
-		wp_cache_set( $cache_key, $styles, $cache_group );
-	}
+	$styles = $merged_data->get_raw_data()['styles'];
 
 	return _wp_array_get( $styles, $path, $styles );
 }